Include plaintext response body in `event`
Presently, XML and JSON response bodies get encoded into the returned
`event` object. These bodies can then be used later in flows (for
filters, or for including data in other destinations). However,
plaintext bodies are not included in the `event` object.

This leads to using success/failure/reason regex in the mappings
themselves to interrogate the response body, as that is presently the
only place where it can be examined. In this manner, one can effectively
"inject" the body (or some portion of it) into the "reason" field. This
also necessarily modifies the "outcome", even if that might not be the
desired effect.

An example would be an enhancement that returns "true" or "false"
depending on whether an email address is a duplicate. In either case,
the "outcome" for the enhancement could be considered "success" --that
is, we only want the "outcome" to be an error or failure if the endpoint
itself failed-- and filters would then want to examine the response body
to take appropriate action. Another example would be an enhancement that
returns a single number as a score. Perhaps we'd like to include that in
a subsequent destination. Presently, since the response body is not
available, the success/failure/reason regex is the only available
mechanism for performing these actions.

This commit adjusts the response handler to include non-XML, non-JSON,
non-HTML responses in a "body" property of the `event` object. This
permits plaintext responses to have the same extensibility as JSON and
XML responses. The commit also adjusts several tests that previously
were expecting just the "outcome" in the `event` object.