Finished initial cookie handler spec
This spec assumes that res.headers['Set-Cookie'] is an array of
strings. Strictly speaking, I don't have visbility into the caller of
response(), so I couldn't really tell you what the argument to the res
parameter of response() looks like. This spec (and the code) will have
to be updated if res.headers['Set-Cookie'] behaves differently than
this.